* Added TerminalDO#OperationalState * Fixed IsFr8OwnTerminal field save * Improved DB migrations for Terminals

+++ Hub/Services/TerminalDiscoveryService.cs
@@ -213,6 +213,7 @@ namespace Hub.Services
                 {
                     terminalDo.ProdUrl = terminal.ProdUrl;
                     terminalDo.ParticipationState = terminal.ParticipationState;
+                    terminalDo.IsFr8OwnTerminal = terminal.IsFr8OwnTerminal;
                 }
             }
 
@@ -284,6 +285,7 @@ namespace Hub.Services
         private async Task<bool> DiscoverInternal(string terminalUrl, bool isUserInitiated)
         {
             Logger.Info($"Starting discovering terminal at '{terminalUrl}'. Reporting about self as the Hub at '{_serverUrl}'.");
+            bool result = false;
 
             try
             {
@@ -325,50 +327,66 @@ namespace Hub.Services
                     { "Fr8HubCallBackUrl", _serverUrl}
                 };
 
-                var terminalRegistrationInfo = await _restfulServiceClient.GetAsync<StandardFr8TerminalCM>(new Uri(terminalUrl + "/discover", UriKind.Absolute), null, headers);
+                StandardFr8TerminalCM terminalRegistrationInfo = null;
+
+                try
+                {
+                    terminalRegistrationInfo = await _restfulServiceClient.GetAsync<StandardFr8TerminalCM>(new Uri(terminalUrl + "/discover", UriKind.Absolute), null, headers);
+                }
+                catch (Exception ex)
+                {
+                    _eventReporter.ActivityTemplateTerminalRegistrationError($"Failed terminal service: {terminalUrl}. Error Message: {ex.Message} ", ex.GetType().Name);
+                    terminalRegistrationInfo = null;
+                }
 
                 if (terminalRegistrationInfo == null)
                 {
+                    // Discovery failed
                     Logger.Error($"Terminal at '{terminalUrl}'  didn't return a valid response to the discovery request.");
-
                     // Set terminal status inactive 
-
-                    return false;
+                    terminalDo.OperationalState = OperationalState.Inactive;
+                    result = false;
                 }
-
-                var activityTemplates = terminalRegistrationInfo.Activities.Select(Mapper.Map<ActivityTemplateDO>).ToList();
-
-                terminalDo.AuthenticationType = terminalRegistrationInfo.Definition.AuthenticationType;
-                terminalDo.Description = terminalRegistrationInfo.Definition.Description;
-                terminalDo.Label = terminalRegistrationInfo.Definition.Label;
-                terminalDo.Name = terminalRegistrationInfo.Definition.Name;
-                terminalDo.Version = terminalRegistrationInfo.Definition.Version;
-                terminalDo.TerminalStatus = terminalRegistrationInfo.Definition.TerminalStatus;
-
-                if (string.IsNullOrWhiteSpace(terminalDo.Label))
+                else
                 {
-                    terminalDo.Label = terminalDo.Name;
-                }
+                    terminalDo.OperationalState = OperationalState.Active;
+                    terminalDo.AuthenticationType = terminalRegistrationInfo.Definition.AuthenticationType;
+                    terminalDo.Description = terminalRegistrationInfo.Definition.Description;
+                    terminalDo.Label = terminalRegistrationInfo.Definition.Label;
+                    terminalDo.Name = terminalRegistrationInfo.Definition.Name;
+                    terminalDo.Version = terminalRegistrationInfo.Definition.Version;
+                    terminalDo.TerminalStatus = terminalRegistrationInfo.Definition.TerminalStatus;
 
+                    if (string.IsNullOrWhiteSpace(terminalDo.Label))
+                    {
+                        terminalDo.Label = terminalDo.Name;
+                    }
+                    result = true;
+                }
                 terminalDo = _terminal.RegisterOrUpdate(terminalDo, isUserInitiated);
 
-                foreach (var curItem in activityTemplates)
+                if (result)
                 {
-                    Logger.Info($"Registering activity '{curItem.Name}' from terminal at '{terminalUrl}'");
-                    try
+                    var activityTemplates = terminalRegistrationInfo.Activities.Select(Mapper.Map<ActivityTemplateDO>).ToList();
+                    foreach (var curItem in activityTemplates)
                     {
-                        curItem.Terminal = terminalDo;
-                        curItem.TerminalId = terminalDo.Id;
+                        Logger.Info($"Registering activity '{curItem.Name}' from terminal at '{terminalUrl}'");
+                        try
+                        {
+                            curItem.Terminal = terminalDo;
+                            curItem.TerminalId = terminalDo.Id;
 
-                        _activityTemplateService.RegisterOrUpdate(curItem);
+                            _activityTemplateService.RegisterOrUpdate(curItem);
+                        }
+                        catch (Exception ex)
+                        {
+                            _eventReporter.ActivityTemplateTerminalRegistrationError($"Failed to register {curItem.Terminal.Name} terminal. Error Message: {ex.Message}", ex.GetType().Name);
+                        }
                     }
-                    catch (Exception ex)
-                    {
-                        _eventReporter.ActivityTemplateTerminalRegistrationError($"Failed to register {curItem.Terminal.Name} terminal. Error Message: {ex.Message}", ex.GetType().Name);
-                    }
+
+                    _activityTemplateService.RemoveInactiveActivities(activityTemplates);
                 }
-
-                _activityTemplateService.RemoveInactiveActivities(activityTemplates);
+                return result;
             }
             catch (Exception ex)
             {
